third_party/WebKit/LayoutTests/fast/dom/shadow/content-pseudo-element-relative-selector-css-text.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
<style id="style1">
.foo::content > div .bar::before { display: block; }
</style>
</head>
<script>
description("Test for cssText of '::distributed()' rule.");
shouldBeEqualToString("document.getElementById('style1').sheet.cssRules.item(0).cssText", ".foo::content > div .bar::before { display: block; }");
</script>
</html>
