third_party/WebKit/LayoutTests/fast/dom/shadow/content-pseudo-element-relative-selector-css-text-expected.txt

Test for cssText of '::content' rule.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.getElementById('style1').sheet.cssRules.item(0).cssText is ".foo::content > div .bar::before { display: block; }"
PASS successfullyParsed is true

TEST COMPLETE
